Basic multiplayer chess functionality
A player can create a new game and share a link with their opponent.
The opponent can then join the game. Long polling is supported so that
the server will only respond to a player once it is their turn.
This lets us avoid endless requests to the server for updates.